Reduce response latency using streams
This commit introduces streams which are used to reduce the latency
between a request for a resource and the start of transferring actual
data.

The issue stems from the fact that large resources need to be cached
completely until the first byte is written to the http response. I
propose streams to solve this issue such that the resource may be cached
and transmitted to the client simultaneously. Stream are concurrent and
thread safe and supports in-memory and file backed data stores.